requirements.txt
discord.py>=2.3.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
PyNaCl>=1.5.0